# src/common/api.yml
openapi: 3.0.1
info:
  title: reseller-api
  description: Boticario Reseller Api
  version: 1.0.0
servers:
  - url: /api/v1
...
paths:
...
  /user/{id}:
    get:
      tags:
        - User
      description: Fetch a specific users
      parameters:
        - name: id
          in: path
          description: The id of the user to retrieve
          required: true
          schema:
            type: string
      responses:
        200:
          description: Returns a specific user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserBody'
    delete:
      tags:
        - User
      description: Delete a specific users
      parameters:
        - name: id
          in: path
          description: The id of the user to be deleted
          required: true
          schema:
            type: string
      responses:
        200:
          description: Delete a specific user
          content: {}
    put:
      tags:
        - User
      description: Update a user
      requestBody:
        description: user model
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserBody'
        required: true
      parameters:
        - name: id
          in: path
          description: The id of the user to update
          required: true
          schema:
            type: string
      responses:
        200:
          description: Returns updated user data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserBody'
...
  /deal/{id}:
    get:
      tags:
        - Deal
      description: Fetch a specific deal
      parameters:
        - name: id
          in: path
          description: The id of the deal to retrieve
          required: true
          schema:
            type: string
      responses:
        200:
          description: Returns a specific deal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DealBody'
    delete:
      tags:
        - Deal
      description: Delete a specific deal
      parameters:
        - name: id
          in: path
          description: The id of the deal to be deleted
          required: true
          schema:
            type: string
      responses:
        200:
          description: Delete a specific deal
          content: {}
...
components:
  schemas:
    DealBody:
      title: deal
      required:
        - cpf
        - value
      type: object
      properties:
        id:
          type: number
        date:
          type: string
          format: date
        cpf:
          type: string
        value:
          type: number
        cashback_percent:
          type: number
        cashback_value:
          type: number
        status:
          type: string
...
    UserBody:
      title: user register request
      required:
        - email
        - username
        - password
        - name
        - user_metadata
      type: object
      properties:
        user_id:
          type: string
        email:
          type: string
        username:
          type: string
        nickname:
          type: string
        picture:
          type: string
        password:
          type: string
        phone_number:
          type: string
        given_name:
          type: string
        family_name:
          type: string
        name:
          type: string
        user_metadata:
          $ref: '#/components/schemas/UserMetadataBody'
        # app_metadata:
        #   type: array
        #   items:
        #     type: 
    UserMetadataBody:
      title: custom user fields
      required:
        - cpf
      type: object
      properties:
        cpf:
          type: string
...